More accurately install color-theme script.

+++ install/install-color-theme.sh
@@ -11,21 +11,30 @@ UTIL_DIR="${CURRENT_DIR}/util"
 WD=`pwd`
 cd $UTIL_DIR
 UTIL_DIR=`pwd`
-cd $WD
 EMACS_HOME=`${UTIL_DIR}/emacs-home.sh`
 EMACS_MODE_HOME=`${UTIL_DIR}/emacs-mode-home.sh`
 
 # run
 
 INSTALLED_MODE_EL_FILE_URL="http://download-mirror.savannah.gnu.org/releases/color-theme/color-theme-6.6.0.zip"
-INSTALLED_MODE_DIR="color-themes"
+INSTALLED_MODE_DIR="color-theme"
 
 cd ${EMACS_MODE_HOME}
 wget ${INSTALLED_MODE_EL_FILE_URL}
 ARCHIVE_FILE=`basename ${INSTALLED_MODE_EL_FILE_URL}`
 unzip -d . ${ARCHIVE_FILE}
-rm ${ARCHIVE_FILE}
-NUMBER_OF_DIRS=`ls -ld color-theme* | wc -l`
-if [ ${NUMBER_OF_DIRS} -eq 1 ]; then
-    mv color-theme* color-theme
+if [ -f ${ARCHIVE_FILE} ]; then
+    rm ${ARCHIVE_FILE}
 fi
+if [ -d ${INSTALLED_MODE_DIR} ]; then
+    rm -rf ${INSTALLED_MODE_DIR}
+fi
+NUMBER_OF_DIRS=`ls -la | grep "color-theme-.*" | wc -l`
+if [ ${NUMBER_OF_DIRS} -eq 1 ]; then
+    UNPACKED_DIR=`ls -la | grep "color-theme-.*" | awk -F\  '{print $9}'`
+    if [ -d ${UNPACKED_DIR} ]; then
+	mv ${UNPACKED_DIR} ${INSTALLED_MODE_DIR}
+    fi
+fi
+
+cd ${WD}
